+++ lib/errbase.rb
@@ -5,7 +5,15 @@ module Errbase
     def report(e, info = {})
       Airbrake.notify(e, info) if defined?(Airbrake)
 
-      Appsignal.send_error(e, info) if defined?(Appsignal)
+      if defined?(Appsignal)
+        if Appsignal::VERSION.to_i >= 3
+          Appsignal.send_error(e) do |transaction|
+            transaction.set_tags(info)
+          end
+        else
+          Appsignal.send_error(e, info)
+        end
+      end
 
       if defined?(Bugsnag)
         Bugsnag.notify(e) do |report|