+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.2 (unreleased)
 
 - Added support for Scout APM
+- Fixed warning with Appsignal
 
 ## 0.2.1 (2021-01-03)